refactor: separate role through operand wrapping
